requirements.txt
click==8.1.3
Flask==2.3.3
gunicorn>=21.0.0
itsdangerous==2.2.0
Jinja2>=3.1.2
MarkupSafe>=2.0.0
Werkzeug>=2.3.3
requests>=2.32.3
Pillow>=11.1.0
pillow-avif-plugin>=1.5.0

python-dotenv>=1.0.0
